Working end-to-end

+++ MaxText/layers/normalizations.py
@@ -42,7 +42,7 @@ class RMSNorm(nn.Module):
         'scale',
         nn.with_logical_partitioning(self.scale_init, self.kernel_axes),
         (features,),
-        jnp.float32,
+        jnp.bfloat16,
     )
 
     scale = jnp.asarray(scale, self.dtype)
